requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
passlib==1.7.4
pyjwt==2.10.1
bcrypt==4.2.1